move SetPermissions to Channels
