+++ Watchman.Discord/Areas/Initialization/Services/MuteRoleInitService.cs
@@ -17,16 +17,24 @@ namespace Watchman.Discord.Areas.Initialization.Services
             this._channelsService = channelsService;
         }
 
-        public async Task InitForServer(DiscordServerContext server)
+        public async Task InitForServerAsync(DiscordServerContext server)
         {
-            var mutedRole = this._usersRolesService.GetRoleByName(UsersRolesService.MUTED_ROLE_NAME, server);
             var changedPermissions = this.CreateChangedPermissions();
+            var mutedRole = this._usersRolesService.GetRoleByName(UsersRolesService.MUTED_ROLE_NAME, server);
             if (mutedRole == null)
             {
                 var createdMutedRole = this.CreateMuteRole(changedPermissions.AllowPermissions);
                 mutedRole = await this.SetRoleToServer(server, createdMutedRole);
+                await Task.Delay(1000); // wait for complete creating a muted role
             }
             await this.SetChannelsPermissions(server, mutedRole, changedPermissions);
+        }
+
+        private ChangedPermissions CreateChangedPermissions()
+        {
+            var noPermissions = new List<Permission>();
+            var denyPermissions = new List<Permission> { Permission.SendMessages, Permission.SendTTSMessages, Permission.CreateInstantInvite, Permission.AddReactions, Permission.ChangeNickname, Permission.Speak };
+            return new ChangedPermissions(noPermissions, denyPermissions);
         }
 
         private NewUserRole CreateMuteRole(ICollection<Permission> permissions)
@@ -34,21 +42,14 @@ namespace Watchman.Discord.Areas.Initialization.Services
             return new NewUserRole(UsersRolesService.MUTED_ROLE_NAME, permissions);
         }
 
-        private async Task<UserRole> SetRoleToServer(DiscordServerContext server, NewUserRole mutedRole)
+        private Task<UserRole> SetRoleToServer(DiscordServerContext server, NewUserRole mutedRole)
         {
-            return await this._usersRolesService.CreateNewRole(server, mutedRole);
+            return this._usersRolesService.CreateNewRole(server, mutedRole);
         }
 
-        private async Task SetChannelsPermissions(DiscordServerContext server, UserRole mutedRole, ChangedPermissions changedPermissions)
+        private Task SetChannelsPermissions(DiscordServerContext server, UserRole mutedRole, ChangedPermissions changedPermissions)
         {
-            await this._channelsService.SetPermissions(server.GetTextChannels(), server, changedPermissions, mutedRole);
-        }
-
-        private ChangedPermissions CreateChangedPermissions()
-        {
-            var noPermissions = new List<Permission>();
-            var denyPermissions = new List<Permission> { Permission.SendMessages, Permission.SendTTSMessages, Permission.CreateInstantInvite };
-            return new ChangedPermissions(noPermissions, denyPermissions);
+            return this._channelsService.SetPermissions(server.GetTextChannels(), server, changedPermissions, mutedRole);
         }
     }
 }
